Process PR feedback.
