+++ Conjugar/ConjugationCell.swift
@@ -50,7 +50,7 @@ class ConjugationCell: UITableViewCell {
         conjugation = personNumber.pronoun + " " + conjugation
       }
       self.conjugation.attributedText = conjugation.conjugatedString
-        self.conjugation.setAccessibilityLabelInSpanish(conjugation)
+      self.conjugation.setAccessibilityLabelInSpanish(conjugation)
     }
   }
