+++ Conjugar/UIViewExtensions.swift
@@ -8,6 +8,7 @@
 
 import UIKit
 
+// MARK: - Animation
 extension UIView {
   func pulsate() {
     let duration: TimeInterval = 0.3
@@ -19,5 +20,20 @@ extension UIView {
         self.transform = CGAffineTransform.identity
       })
     })
+  }
+}
+
+// MARK: - Accessibility
+extension UIView {
+  func setAccessibilityLabelInSpanish(_ label: String, region: String = Current.settings.region.accent) {
+    setAccessibilityLabel(label, spokenInLanguage: "es_\(region)")
+  }
+  
+  private func setAccessibilityLabel(_ label: String, spokenInLanguage languageCode: String) {
+    let attributes: [NSAttributedString.Key : Any] = [
+      .accessibilitySpeechLanguage: languageCode
+    ]
+    let attributedLabel = NSAttributedString(string: label, attributes: attributes)
+    accessibilityAttributedLabel = attributedLabel
   }
 }
